requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
billiard==4.2.4
brotli==1.2.0
celery==5.6.2
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorlog==6.10.1
constantly==23.10.4
coverage==7.10.7
cryptography==46.0.2
daphne==4.2.1
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
drf-yasg==1.21.11
factory_boy==3.3.3
Faker==37.8.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
Markdown==3.9
MarkupSafe==3.0.3
model-bakery==1.20.5
msgpack==1.1.2
openai==2.6.0
packaging==25.0
paypalrestsdk==1.13.3
paytmchecksum==1.7.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyotp==2.9.0
pytest==8.4.2
pytest-django==4.11.1
pytest-factoryboy==2.8.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
razorpay==2.0.0
redis==7.3.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sendgrid==6.12.5
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==13.0.1
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.3
whitenoise==6.11.0
zope.interface==8.0.1
